Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> I am happy to continue the theoretical discussion, however in the meantime, do you agree that CL is small incremental improvement (or at least does not cause harm), and thus can be submitted?

Please at least document (1) that this cheats on atomicity, and (2) document the behavior when you read odd version numbers, to the extent your code will rely on it. That relies on very subtle properties of the code that are not otherwise apparent. If you do want to rely on this, please also make at least the link fields atomic. Memory_order_relaxed accesses should be fine.

One reason I'm hesitant to take changes like this is that they will all need to be redone if we ever get TSAN support. TSAN will complain about the racing accesses, and it will be correct.

I'm currently in a week-long WG21 meeting. Sorry about the slow response.

Patch-set: 5
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>