Update patch set 6
Patch Set 6: Presubmit-Verified+1

=== Finished presubmit run: L59500000152780324 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/627227?revision=6&workplanId=L59500000152780324




TreeHugger finished with: 15 passed.

WARNING: The tip-of-tree(4655872) for aosp-master/aosp_arm64-eng may be broken.
The presubmit build passes with the last-known-good-build(4655728), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4655872) for aosp-master/aosp_x86-eng may be broken.
The presubmit build passes with the last-known-good-build(4655637), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






















This is an awesome step forward. Great job!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 6
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0
Label: Presubmit-Verified=+1