Try to prevent stdout/stderr interleaving in test 1940
In host configurations test 1940 at one point has some std-err output
from native code. This is filtered out but if it ends up interleaving
with the std-out output we will end up missing some of the output and
incorrectly failing the test.

Put in some 'flush' calls around where the printing takes place to
make this not happen.

Test: ./test.py --host -j50
Change-Id: Ic42816beeea2f04046b5bb88e17a1f6bde19a58e