Try to prevent System.err/System.out interleaving in test 1940
In host configurations test 1940 at one point has some System.err
output from an ExceptionDescribe (debugger.cc:4358). This is filtered
out but if it ends up interleaving with the System.out output we will
end up missing some of the output and incorrectly failing the test.

Put in some 'flush' calls around where the printing takes place to
make this interleaving not happen.

Test: ./test.py --host -j50
Bug: 70988713
Change-Id: I41e743c42e63c0dcb6d32473c31f391983819d32