Allow hidden API access from system libraries
Libraries like MediaSession2 and RemoteDisplay provide a library
that an app loads into its process and which accesses internal
APIs on the app's behalf, without exposing the internals to the app.
These libraries are considered part of the framework, but were not
exempt from hidden API checks because they are not loaded with the
boot strap class loader.

This patch adds a new flag to DexFile class which the constructor
sets to true of the canonical location of the newly loaded dex file
starts with "${ANDROID_ROOT}/framework/". Hidden API enforcement
then checks this flag when determining whether the caller of
a hidden class member is allowed to access it or not.

Bug: 64382372
Bug: 76138670
Test: art_dex_file_loader_test gtest
Test: art/test.py --gtest
Change-Id: If062bd668d7ba494bbb7b828e40932748d173b9a