Log an event on all hidden API accesses.
The new event consists of:
- The type of access (reflection, JNI, etc.)
- The action taken (warn or deny)
- The type of member accessed (field or method)
- Name of the class which defined the member accessed
- The name of the member

The type of the member is not included, as it can be inferred from the
class and member name. The fully qualified member name is also not included
to avoid the overhead of building the string.

Similarly, the package name, version, etc. are not included as they can
be inferred from the context when analyzing the event log.

Test: $ m
Test: $ adb lolcat -b events | grep art_hidden_api_access

Sample output:
2612  2612 I art_hidden_api_access: [0,0,Landroid/app/Activity;,mWindowManager]
2612  2612 I art_hidden_api_access: [0,0,Landroid/app/Activity;,mWindow]
2700  2700 I art_hidden_api_access: [2,1,Ljava/lang/System;,arraycopy]
(Timestamps have been elided)

Change-Id: I471dbeb3ba3bb1a1ea0451ff7a8749daf0855317