Update patch set 1
Patch Set 1:

> > Please factor this code differently. For monitor logging, we decided on a separate file for target builds.
> 
> can you give me a code pointer? (I've never done this before, so just guessing)

I'm arguing it's time to stop pretending there's no overhead to hidden-api and have everything all-inlined header-only. The check itself may be, but the reaction to a check is no longer so.

I'm suggesting you add hidden_api .cc files, and the logging implementation is only used for the target build. You can take monitor_android.cc as an example.

Patch-set: 1