Ensure that OSR still is possible with jvmti
We were previously disabling OSR by always claiming to be interested
in every method. This could cause slowdown on some methods. To fix
this we correctly only claim to be interested in methods if we might
hit breakpoints from an invoke, have modified locals, or have forced
the function to the interpreter.

There was also a minor bug in the instrumentation removal code that
caused prebuilt compiled code to be used when it should have been
ignored.

In the future we should get the granularity down to single frames but
that is currently not possible.

Test: ./test.py --host -j50 --all -t 993
Test: ./test.py --host -j50
Test: am start --attach-agent -n com.example.android.displayingbitmaps/.ui.ImageGridActivity
      Run blur filter.
Bug: 76226464

Change-Id: I6a7aa0c6353372aebe01613b66841543614f3054