Ensure that OSR still is possible with jvmti
We were previously disabling OSR by always claiming to be interested
in every method. This could cause slowdown on some methods. To fix
this we correctly only claim to be interested in methods if we have
breakpoints or might have modified the locals of some frame.

In the future we should get the granularity down to single frames but
that is currently not possible.

TODO Still test the studio stuff.
Test: ./test.py --host -j50
Test: am start --attach-agent -n com.example.android.displayingbitmaps/.ui.ImageGridActivity
      Run blur filter.
Bug: 76226464

Change-Id: I6a7aa0c6353372aebe01613b66841543614f3054