Update patch set 1
Patch Set 1:

> > Patch Set 1:
 > >
 > > > Patch Set 1:
 > > >
 > > > How frequently does this trigger in practice today? eventlog is
 > down to a handful of hours already, I don't really want to lose
 > more time out of this log since we don't have the log spam
 > prevention that the main log does.
 > >
 > > I wasn't aware of this concern, perhaps this suggests this go to
 > a WTF instead since those will be collected elsewhere?
 > 
 > Thinking about this more, if we are only collecting dark grey and
 > black list data, the amount should be very small. its the light
 > grey list that is spammy.

Confirmed. If spam is an issue, then dark grey and black shouldn't have much logs (with the caveat that a blacklist access will always be logged, so if the same app / PID tries to access it twice, it will be logged twice).

If we only only log for dark grey / black, there won't be any performance concerns. Andreas if you still have some, please raise the issue in our meetings.

Patch-set: 1
CC: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>