Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> How frequently does this trigger in practice today? eventlog is down to a handful of hours already, I don't really want to lose more time out of this log since we don't have the log spam prevention that the main log does.

Not sure, it will depend on what apps you are using; I'll try to get an idea. Is there an alternative way to get hold of this sort of data?

I can add a config flag, and only enable the logging on devices where we actually plan to use it, to limit the impact if that would help.

How big is the event log buffer?

Patch-set: 1
Reviewer: Gerrit User 1010681 <1010681@85c56323-6fa9-3386-8a01-6480fb634889>