Log an event on hidden API accesses.
The new event consists of:
- The type of access (reflection, JNI, etc.)
- The action taken (warn or deny)
- The type of member accessed (field or method)
- Name of the class which defined the member accessed
- The name of the member
- The type signature of the member (type of field, or method signature)

The fully qualified member name is also not included to avoid the overhead
of building the string. It can be build from the information included.

Similarly, the package name, version, etc. are not included as they can
be inferred from the context when analyzing the event log.

The event is sampled, according to a sampling rate that can be set by a
configuration option, to reduce log spam.

Test: $ m
Test: $ adb shell settings put global hidden_api_access_log_sampling_rate 100
Test: $ adb lolcat -b events | grep art_hidden_api_access

Sample output:
 2701  2701 I art_hidden_api_access: [0,1,Landroid/app/Activity;,mCalled,Z]
 2701  2701 I art_hidden_api_access: [0,3,Landroid/app/Activity;,mDoReportFullyDrawn,Z]
(Timestamps have been elided)

Change-Id: I471dbeb3ba3bb1a1ea0451ff7a8749daf0855317