Update patch set 3
Patch Set 3:

> > These test are closely related, they can be merged easily and are
 > > unlikely to time out. Most other tests do not have these
 > > attributes.
 > >
 > > The time needed to run an individual test is actually quite high
 > at
 > > the moment, so fewer tests means faster testing.
 > 
 > Do you plan to refactor more tests too?
 > 
 > I am in full agreement that this is a good idea if testing overhead
 > is indeed becoming an issue, but I am just very surprised that
 > while it is evident I am very actively working on min/max/abs, such
 > a change is dropped on me. Not once during any of the prior reviews
 > was this issue raised.
 > 
 > Just a minor suggestion for improving LON-MTV collaboration. I am
 > only one email away....
 > 
 > But ok, proceed with this change.

Isn't this CL a means for discussion? IIRC it's only a suggestion based on having the tests failing in the same way in one of Vladimir's run.

Note that I'm in general not in favor of merging multiple tests into one. But won't object if it makes sense to you and Vladimir.

Patch-set: 3
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>