WIP: Test to analyze overhead of memory protection based GC
Test program to analyze the overhead of memory protection based
compaction. The program creates a page worth of 8-word long objects,
randomly spread over N source pages and then when a read access on
a destination page faults, the handler moves the objects to the
destination page and enable access. This is performed in for M
iterations and average time is reported. Hardware caches and TLBs
are flushed in each iteration.

Test: mma test-art-target-gtest-segfault_cost_test64
Change-Id: I78db26a0b0d737709ce2a2f0342cd2e2d838e294