Update patch set 4
Patch Set 4:

I uploaded a new version of go/aog/637929 that might fix the dead-variable elimination problem, but not the earlier gcstress issue (except that it's currently based on a tree version with vmarko's CL). It conflicts with this CL. Do you want to leave this one reverted, and I'll include the equivalent of this CL there? I'm open to other suggestions, but the tree will eventually have to hold still long enough ..

Patch-set: 4