Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

Are we convinced that the skipped block code is actually necessary/useful?  IIUC, all of this code is only exercised if another thread beats us to copying an object between the time we first discover that it needs to be copied and we finish? The probability of that seems tiny. And the down side of not handling it specially seems small: We just allocate another garbage block, which will hopefully be reused after the next GC.

I would be inclined to remove all the memory_order_seq_cst arguments to infrequent counter updates here. Since that's the default anyway, I think they falsely imply that we've thought about and the choice of seq_cst is deliberate here. I suspect they could all be relaxed, but that would require more scrutiny, which probably isn't worth the effort. But we can do that in a separate pass.

The objects_moved_ and bytes_moved_ counters here also need to be revisited in a separate CL. That's already on oth's list. They do look performance-critical to me. Memory order seems inconsistent, and they look like a cache contention disaster. It looks to me like every running thread will be beating on these at the same time.

In any case, this CL is an improvement, and the restructuring is good even if the counter code eventually goes away.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2