Revert^4 "Ensure that OSR still is possible with jvmti"
This reverts commit ce2836aaf9b04a0ad23739b24adc4437b6443bd3.
This unreverts commit b9ad26d1ed9146b89555d4333021f44eeb831f05.

Reason for revert: Fixed issue with test 1935 causing flakiness.

It was possible for the (rather small) jit-code-cache to fill up
before we call ensureJitCompiled in test 1935. This call as a
side-effect turns off jit garbage collection which means are never
able to compile the method and end up failing the test.

Change-Id: I9aaf9eaf6d240637567359cd72d4f172e5326511
Test: ./test.py --host -j50 --all -t 993
Test: ./test.py --host
Test: while ./test/run-test --host --jit 1935; do; done
Test: while ./test/run-test --host --jit --jvmti-redefine-stress 1935; do; done
Test: am start --attach-agent -n com.example.android.displayingbitmaps/.ui.ImageGridActivity
      Run blur filter.
Bug: 76226464