Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

You are right.
The specific situation I wanted to cover was present only in the previous revision of 616-cha-unloading.

If you remember, there was a println() between the moment when a classloader is deleted and when a problematic arena is trying to be reused (reuseArenaOfMethod()). That println() triggered DefineClass("BufferedWriter"). In case of 8.1, it led to the problem that the problematic arena was reused under boot classloader. Since reuseArenaOfMethod() doesn't try to look for arena being reused by an existing classloader, the method could not complete its purpose and the test failed when it shouldn't.

When you updated the test with a3cd124, you luckily excluded such "false positive" behaviour by removing println(). To prevent the possibility of such situations in future in case something will be changed, I suggested this patch.

Patch-set: 1
Reviewer: Gerrit User 1121591 <1121591@85c56323-6fa9-3386-8a01-6480fb634889>