Load-Store Elimination: Branching Support.
For the full description, see the design doc linked in b/77906240.
This CL is Phase 1 of that plan.

Quick version:

Load-Store Elimination (LSE) steps through every instruction in every
block, keeping track of a virtual heap. If any loads or stores are
redundant, it eliminates them.

At every block, before visiting the instructions, it merges the virtual
heaps for all its predecessors. If different predecessors have different
values for a particular heap location, LSE gives up and sets the
location to 'unknown'. This CL instead creates a Phi at the begining of
the block.

As a result, some loads will now be replaced by Phis.

Golem patch run was perf neutral.

Bug: b/77906240
Test: m test-art-host, target

Change-Id: I3a9f6e526a59cbf969853acbb68d40df2c875c31