Update patch set 6
Patch Set 6: Presubmit-Verified+1

=== Finished presubmit run: L39600000162378051 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/662100?revision=6&workplanId=L39600000162378051




TreeHugger finished with: 16 passed (6 new).












WARNING: The tip-of-tree(4715189) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4714964), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






WARNING: The tip-of-tree(4715179) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4714455), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4715179) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4714964), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






I'm your biggest fan!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 6
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1