Make adbconnection try to start the JIT if it is not running.
The new adbconnection jdwp-provider forces processes to throw out
compiled code when a debugger attaches. This is normally not a problem
since the JIT will pick up the slack in short order. If one is
debugging system_server (or a few other similar system services)
performance will be permanently degraded since these processes run
with the jit disabled.

In order to improve the experience of debugging these processes we
make adbconnection attempt to start the jit if it is not currently
running. In order for this to work the user will need to run
'stop; setenforce 0; start' to disable selinux prior to attaching the
debugger.

Bug: 78119634

Test: adb root &&
      adb shell stop &&
      adb shell setenforce 0 &&
      adb forward tcp:12345 jdwp:`adb shell pidof system_server` &&
      jdb -attach localhost:12345;
      <Do things with debugger and device>

Change-Id: Iffb382460a1dbcd055d533a44367c4ccb3830f10