Use spin loop instead of sched_yield to spin
Inflating a lock used to call sched_yield 50 times in a loop on
contention before deciding to inflate. Let's use a spin loop instead
so that we can grab the lock faster for very short critical sections
and so that the locking thread can avoid giving up its timeslice until
it decides it wants to sleep on the futex.

Test: ART host tests
Bug: 78081829
Change-Id: Ic4141b55f7fe4594e738da36f32d14147977df3d