ARM: Use rMR for Baker RB introspection marking.
The marking register is known to be 1 when entering the
introspection marking entrypoint, so we can clobber it and
reload the 1 before returning instead of reserving a special
register r4. The immediate benefits are minor, see below,
but this shall allow further improvements, for example we
could try to change rMR to r4 which would reduce code size
of every marking register check by 2 bytes.

ARM boot image (boot*.oat) size in aosp_taimen-userdebug:
  - before: 17861724
  - after: 17858088 (-3636)

Test: Pixel 2 XL boots.
Test: testrunner.py --target --optimizing --32
Bug: 36141117
Change-Id: I0f625dec3a6b3ee1786f7e5f4377be42b9bc37d3