Do not verify patched art files for heap poisoning.
We do not have a way to correctly distinguish poisoned
references from native pointers living on the heap (for
example ArtMethod*s stored in Class objects). That would
require knowing the ranges of all boot image object spaces,
not just the current one. And the boot image address range
would have to avoid crossing the 2GiB address, so that the
same value cannot be valid as both a poisoned object
reference and native pointer.

With no way to verify the diggest, just accept the file
anyway as the heap poisoning is a testing configuration that
should never ship to users.

Test: m ART_HEAP_POISONING=true
Test: Pixel 2 XL boots with heap poisoning.
Bug: 66697305
Change-Id: I619d2755925da08703a7671454c5394a7ee0158c