Update patch set 9
Patch Set 9:

Discussed with Vladimir and Aart: this plan will be taken over in the near future.

Thank you for the many days of review; I hope this CL will serve as a good base for the next engineer to work on this.

Patch-set: 9
Reviewer: Gerrit User 1272179 <1272179@85c56323-6fa9-3386-8a01-6480fb634889>