Update patch set 2
Patch Set 2: Presubmit-Verified-1

=== Finished presubmit run: L84500000165981327 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/670203?revision=2&workplanId=L84500000165981327




TreeHugger finished with: 15 passed, 2 failed.

WARNING: The tip-of-tree(4737934) for aosp-master/aosp_arm64-eng may be broken.
The presubmit build passes with the last-known-good-build(4737729), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



FAIL: Failed to build aosp-master/aosp_x86-eng/4737934 (ToT). See https://android-build.googleplex.com/builds/pending/P5825106/aosp_x86-eng/latest. 
Note: This is a checkbuild (go/checkbuild). It will take longer and could uncover errors not found in other builds.

WARNING: The tip-of-tree(4737996) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4737712), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4737996) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4737712), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




FAIL: Failed to build git_master/sailfish-userdebug_fastbuild3_linux/4737996 (ToT). See https://android-build.googleplex.com/builds/pending/P5825096/sailfish-userdebug_fastbuild3_linux/latest. 
Note: This is a checkbuild (go/checkbuild). It will take longer and could uncover errors not found in other builds.

WARNING: The tip-of-tree(4737996) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4737712), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4737996) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4737712), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 2
Tag: autogenerated:TreeHugger:Verify-Fail
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=-1