Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Eh, was worth a shot.
> > > > 
> > > > ART does use SafeMap and its wrappers in a bunch of places. It might be nice to move some of them to a data structure like this one day.
> > > 
> > > I am not against this change per se. The HashMap could still be smaller when empty than the SafeMap, which means an overall win.
> > > 
> > > I am against a CL with no data what the tradeoffs are.
> > 
> > Is there a specific benchmark that exercises this code path?
> 
> Any code that is being verified and has locks. :-)
> 
> You can derive the interesting data by hand, though. That's what Mathieu did when introducing the hash set. That's what vmarko does when moving away from std::map. Just tell me why it should be better for regular cases (one or two locked registers, usually never more than five).

It's probably not better when kMinBuckets is 1000. :-) (Below a certain size, we should probably not bother with hashing at all, instead doing a linear scan.)

I still think it'd be worthwhile to move toward unordered dictionaries generally (once HashMap works better for small numbers of objects), but I worry that it'd be hard to justify each specific substitution because the effect of each change won't be particularly large and we're probably touch parts of the runtime that are hard to isolate for microbenchmarks. Still, in a broad sense, we'll win overall with more locality and less pointer chasing.

Patch-set: 2
Reviewer: Gerrit User 1130394 <1130394@85c56323-6fa9-3386-8a01-6480fb634889>