Update patch set 2
Patch Set 2:

(2 comments)

I am not that fond  of this change. 

It is nice to know that odex & art files are always loaded together and not at all. Conditionally loading the art file only adds complexity and uncertainty because what this is saying is: it's ok to load the oat file without duplicate check but it's not ok to do the same thing for the art image. 

Things can go wrong either way...

We only create an app image for base apks and for the http library. I suggest we either:
1) compile the http without '&' and thus always expect it to the the first element of the class loader. Seem reasonable?
2) don't generate an app image for it.

Patch-set: 2
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>