Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > I am not that fond  of this change.
> 
> I wash my hands. I found you the bug, you fix it.
> 
>  > It is nice to know that odex & art files are always loaded together
>  > and not at all. Conditionally loading the art file only adds
>  > complexity and uncertainty because what this is saying is: it's ok
>  > to load the oat file without duplicate check but it's not ok to do
>  > the same thing for the art image.
>  > 
>  > Things can go wrong either way...
> 
> We use the oat files even when it's dangerous for performance reasons. I consider this wrong and do not really want to touch this with a 10m pole.
> 
>  > We only create an app image for base apks and for the http library.
>  > I suggest we either:
>  > 1) compile the http without '&' and thus always expect it to the
>  > the first element of the class loader. Seem reasonable?
>  > 2) don't generate an app image for it.
> 
> This is not about the http library, it's about all shared libraries and all weird class loaders. The regression test does not use any library at all.

We already have a problem with the odex files and I don't think that the approach here is making things betters. It just (ad-hoc) fixes a problem with the app image that's happening because of framework bugs. 

I know we don't reject oat files because of perf reasons. And we all agree is bad. We long stated that the use of '&' is bad. 

But I still think this is just a partial solution that adds more complexity.

I'll leave it to Mathieu to decide if he wants these extra checks or improve the image loading code (b/79200502). Either way we should only do this for '&' loads and if the oat file is not the first one in the class path.

Patch-set: 2
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>