WIP: Make method entry/exit jvmti tracing use stub-instrumentation
We were using interpreter instrumentation to implement method
entry/exit tracking in JVMTI. This change replaces that with
stub-instrumentation.

Some minor bugs in the instrumentation stubs needed to be fixed as
well. We were incorrectly calling unwind-instrumentation functions and
would not use JIT code for the tail call.

Currently have an issue that we will sometimes compile the same method
multiple times.

Bug: TBD
Test: ./test.py --host
Change-Id: I57e0998078aafff1318c996d902c1d7ea693e0fe