Move and rewrite bit loading/storing methods.
Move bit loading/storing methods to BitMemoryRegion and rewrite them.

Enforce natural alignment of the data pointer in BitMemoryRegion.
This probably would not be reasonable in MemoryRegion, but it is
fine here since the BitMemoryRegion already has extra bit offset.
The alignment makes it possible to simplify and optimize the methods.

This makes the stackmap reading code 33% faster.
(measured by the time needed to verify all stackmap fields on ARM).

Test: m -j40 test-art-host-gtest
Change-Id: I00be8052969a6056b262df4cd2066ffd86043196