ART-tests: remove DX-dependency from 166-bad-interface-super
This tests is creating an illegal class file which compiles with DX
but D8 rejects it.

To make it pass with D8 this CL adds the precompiled, smali versions of
the jasmin sources. The build script will use either the smali or the
jasmin sources depending on the --host/--target/--jvm flag.

Bug: 65168732
Test: art/test.py --host -r -t 166-bad-interface-super
Change-Id: I58e69fa9a4aa08946b7e57001f5e2a2a2bc07b0f