Make static libtistress version
It turned out to be difficult to actually run the libtistress on
actual apps due to the fact it linked shared to a lot of libraries.
This adds libtistresss and libtistressds targets which build static
versions of libtistress both without and with debug checks.

This also adds a simple Agent_OnAttach entrypoint for the agent. No
special processing is performed. The agent is setup as though it were
running OnLoad.

Test: m -j50 libtistressds
      adb push $OUT/data/nativetest64/art/arm64/libtistressds.so /data/local/tmp
      adb shell setenforce 0
      adb shell am start-activity \
        --attach-agent /data/local/tmp/libtistressds.so=jvmti-stress,trace \
        com.antonioleiva.bandhookkotlin/.ui.screens.main.MainActivity

Change-Id: I75d3a81011864c62cde785fd7351c59dbd269237