Add ti-fast jvmti agent.
It is often useful to get a speed-of-light estimate for how heavy
watching various JVMTI is. This was difficult to measure without
creating a custom agent to exercise the feature. This adds a new
ti-fast agent which can be used to run a program which has an agent
that does nothing watching various events.

Test: m -j50 libtifast
      adb push $OUT/system/lib64/libtifast.so /data/local/tmp
      adb shell setenforce 0
      adb shell am start-activity \
        --attach-agent /data/local/tmp/libtifast.so=MethodEntry,MethodExit \
        com.antonioleiva.bandhookkotlin/.ui.screens.main.MainActivity

Change-Id: Id97135bcae55f9a943108fd935031fd887fb913a