ART-tests: Remove DX dependency from 450-checker-types.
One of the test cases (testArraySimpleRemove) tests whether the ART
optimizer can remove a trivial check-cast. The test fails with D8
because D8 already removes the check-cast.

This CL, in addition to enabling D8 for the test, replaces the java
source for that single test case with smali so the check-cast will
be there for the ART optimizer.

Test: art/test.py -r --host -t 450
Bug: 65168732
Change-Id: I6f319c7b9d8f2c72f9c739890acd6499f7ffcacd