Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Why is D8 validating the semantics in this case? I don't think a dexer should be in the business of doing this sort of bytecode validation.

Vladimir, D8 does try to be liberal in what it accepts but not in what it outputs, we don't want to output something which is known to be invalid. Is it acceptable if we see it this way?

Patch-set: 6
Reviewer: Gerrit User 1263609 <1263609@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0