Update patch set 1
Patch Set 1: Code-Review+2

This is probably okay as long as clang doesn't care about jumping to a label that is not a function. It does feel a bit weird to do this.

Another way to do this would be to change all of these labels to .L<label_name> but I don't know how the labels are being used, and this might not work.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2