Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> (1 comment)
> 
> I liked the table-based d

It's hard/ugly to do that in-conjunction with the macro's requested by ngeoffray.

How hard is it to get the compiler to allocate the result and perform the appropriate move? We know at compile time what the return type will be. Effectively we'd like to return a jvalue and have the compiler emit the appropriate opaque cast to the expected type.

Patch-set: 8
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0