ART-tests: Remove DX-dependency from 551-checker-shifter-operand
The merging of shift and add operations into DataProcWithShifterOp in the
test case $opt$validateShiftLongAsserts was dependent on register-
allocation patterns. With D8 the merge did not take place if the the
result of the shift was not overwritten by the subsequent add.

This CL separates the intermittent method calls and the arithmetic
operations so the environmental uses of the shift result does not
block merging. Also, D8 is enabled.

Test: art/test.py -r -b --target -t 551-checker-shifter-operand
      Tested locally with --host and --gcstress.

Bug: 65168732
Change-Id: I3a8e71f3530211878879faae068acb7c7957ae19