Add option to view all instances of a class.
Not just direct instances, but also instances of subclasses of the
class.

The objects view now includes a description of what objects are being
shown at the top of the page, such as:

    Site:           ROOT -
    Class:          android.os.Binder
    Subclasses:     excluded (switch to included)
    Heap:           any (switch to app, image, zygote)
    Count:          17,424

Where 'included', 'app', 'image', and 'zygote' are links the user can
use to change the filter for selected instances.

Test: atest ahat-tests --host, with new test added.
Test: manually launch ahat, view objects page and modify filters.
Bug: 109789538
Bug: 111439037

Change-Id: Iedb7d9cf9252335f24208e7f5ba052bf3845e58c