Simplify trampoline instrumentation event dispatch
Previously we would only send instrumentation events from the
trampoline if we are jumping to non-interpreter code. This is fine but
it makes it difficult to add support for jumping to JIT code in the
trampoline because we might have a suspend in the event handler.

To fix this issue we rewrote the trampolines and the interpreter
bridge to allow us to always send the events and tell the interpreter
to skip sending the events a second time.

This also allows us to simplify the ASM part of the instrumentation
trampoline since it no longer needs to deal with throwing exceptions.

Test: ./test/testrunner/testrunner.py --host --runtime-option=-Xplugin:libtracefast-trampolined.so

Change-Id: I658c2daa18fd5eaa1d1375a5709d45934b4560fb